Fix infobars being hidden right after they appear
The PR fixes infobars intermittently hidden right after being shown.
The reason was that we had fragile dismiss logic on navigation start
which led to race conditions basing on if all possible navigations
started before we show the infobar. Infobars are already dismissed
by any touch outside of them, so the fix is we remove navigation based
hiding for infobars.